Added some plotting features.
Now the linkage tree for training data is returned by analyse_single_configuration, and this is used to display a dendrogram of the training dataset. Also, a list of lists is used to record the information values at the point where n_clusts=n_stim_patterns, and this can be used to better compare, for example, the effect of thresholding vs network scale.