KL divergence: first attempt at an implementation.
I'm trying to calculate the KL divergence between the (flat) 'prior' input distribution and the 'decoded' distribution (that is, the distribution over the output space when it has the same dimensionality of the input space). PLease note that this makes sense without specifying an explicit output->decoded input relationship just because the original input distribution is flat.