Trying to delegate archival operations in their own Archive class, and to incorporate path definitions directly in ParamSpacePoint.
Note that this calls for the base ParamSpacePoint class to be written in pure python, so that we can use it in the simulate.py script which is meant to be run in Jython, while at the same time subclassing it from the CPython side of things (master_script, compression, analysis and visualisation steps). To this end, I created a new submodule called 'pure´, which is supposed to replace paths.py.

Not yet functional (more or less everything is broken at this stage).