Job management: remove dependency on queue mgr running on login nodes.
Only use job arrays and inter-job dependencies to regulate the
workflow, as opposed to having my own deamon for queue management
running on a login nodes.

Pros:
- compression jobs now can start as soon as all of the relevant
simulations are done, rather than waiting for all simulations (this
should help with decreasing the stress on the filesystem)
- I don't risk leaving the workflow in a very messy, halfway-done
state if something goes wrong (like if the queue management daemon
crashed)
- I'm actually complying to the guidelines on how to use the cluster
- this is needed if at some point I want to try running the workflow
on Legion, or elsewhere.

Cons:
- I can't limit any more how many processors I take up on the cluster,
and I'll have to rely on the scheduler to avoid clogging it up and
allowing other people in the lab to use it.