Store simulation files in tar archive. Improve use of temp dirs.
Tar archival and hdf5 compression happen in the appopriate temp
dir, which is automatically defined, created and deleted with a
context manager depending on the machine we're running on. Data files
are transferred to their final location only when they're
complete. This should grant a significative performance improvement
especially on Legion, where temporary directories are local to compute
nodes and performance of the scratch filesystem is poor when dealing
with large numbers of files.