SGE: increase requested resources on Matlem.
In general, because I'm more or less the only user on Matlem I don't
have a cost associated with asking for more time. Memory, on the other
hand, affects how many jobs I can run concurrently, but that's not
really an issue for compression and analysis jobs as only a few (tens,
at most) of them are meant to be running at any given time, as the
data they depend on gets generated.