Always use $TMPDIR as scratch space when defined.
Rationale: Jython places some foreign function interface
files (jffi*.tmp) in the location defined by java.io.tmpdir, but this
does not seem to be configurable by an environment variable on
Linux. This is a problem for BLCR, as the default value for
java.io.tmpdir is "/tmp", which is not checkpointed, and these files
are needed to restart failed jobs. The solution I found was to add a
-Djava.io.tmpdir=$TMPDIR option to nC.sh (or rather, to do that only
when $TMPDIR is defined), and to rely on the fact that $TMPDIR should
point to a location which is being checkpointed. Then, since because
of this I already have to specify the location of the
local (checkpointed) scratch space in the submission scripts, I
restructured the ClusterSystem context manager to take advantage of
this. Now the context manager is (almost) system agnostic, as it
doesn't need to be told which machine the code is running on. This, in
turn, has the nice side effect of removing the requirement for the
system information to be passed on to the simulation, compression and
analysis scripts.