Analysis: switch from mini-batch k-means to k-means proper.
The (non mini-) batch implementation is slower (it takes about 5 hours
for a node on Matlem to run the analysis for 30*1024 data points in
roughly 500 dimensions) but because it's repeated several (10) times
with different random initialisations it is much less likely to get
stuck in a local minimum for inertia. Moreover, the performance of the
algorithm itself is inherently superior as the mini-batch
implementation is only supposed to give an (albeit good) approximation
to what the batch implementation does. So, provided we can afford it
from the standpoint of the computational resources, we should go
batch.